feat: 이름 변경 및 comment entity 추가
