+++ src/main/java/codesquad/domain/entity/Comment.java
@@ -1,0 +1,4 @@
+package codesquad.domain.entity;
+
+public record Comment(String userid, String postid, String contents) {
+}
